Refactor: Move route configuration to a separate file, linting
