+++ src/components/map/GoogleMap.svelte
@@ -14,13 +14,12 @@
 	import LocationButton from '$lib/LocationButton/LocationButton.svelte';
 	import StopMarker from './StopMarker.svelte';
 	import RouteMap from './RouteMap.svelte';
+	import { faBus } from '@fortawesome/free-solid-svg-icons';
 	import {
-		faBus,
-		faFerry,
-		faTrainSubway,
-		faTrain,
-		faCableCar
-	} from '@fortawesome/free-solid-svg-icons';
+		RouteType,
+		routePriorities,
+		prioritizedRouteTypeForDisplay
+	} from '../../config/routeConfig';
 
 	export let selectedTrip = null;
 	export let selectedRoute = null;
@@ -38,47 +37,6 @@
 	let markers = [];
 	let allStops = [];
 	let routeReference = [];
-
-	const RouteType = {
-		LIGHT_RAIL: 0,
-		SUBWAY: 1,
-		RAIL: 2,
-		BUS: 3,
-		FERRY: 4,
-		CABLE_CAR: 5,
-		GONDOLA: 6,
-		FUNICULAR: 7,
-		UNKNOWN: 999
-	};
-
-	const routePriorities = [
-		RouteType.FERRY,
-		RouteType.LIGHT_RAIL,
-		RouteType.SUBWAY,
-		RouteType.RAIL,
-		RouteType.BUS,
-		RouteType.CABLE_CAR,
-		RouteType.GONDOLA,
-		RouteType.FUNICULAR,
-		RouteType.UNKNOWN
-	];
-
-	const prioritizedRouteTypeForDisplay = (routeType) => {
-		switch (routeType) {
-			case RouteType.FERRY:
-				return faFerry;
-			case RouteType.LIGHT_RAIL:
-			case RouteType.SUBWAY:
-			case RouteType.CABLE_CAR:
-				return faTrainSubway;
-			case RouteType.RAIL:
-				return faTrain;
-			case RouteType.GONDOLA:
-				return faCableCar;
-			default:
-				return faBus;
-		}
-	};
 
 	async function loadStopsForLocation(lat, lng) {
 		const response = await fetch(`/api/oba/stops-for-location?lat=${lat}&lng=${lng}`);
